# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: string;
  name: string;
  category: string;
  description: string;
  price: number;
  currency: string;
  image: string;
  material: string;
  featured: boolean;
  /** @nullable */
  badge: string | null;
  /** @nullable */
  stripePriceId: string | null;
}

export interface StoreSummary {
  featured: Product[];
  categories: string[];
  productCount: number;
}

export interface CheckoutLineItem {
  productId: string;
  /**
     * @minimum 1
     * @maximum 10
     */
  quantity: number;
}

export interface CheckoutInput {
  /**
     * @minItems 1
     * @maxItems 20
     */
  items: CheckoutLineItem[];
}

export interface CheckoutSession {
  sessionId: string;
  url: string;
}

export type ListProductsParams = {
category?: string;
featured?: boolean;
};