# lib/api-zod/src/generated/types/checkoutInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CheckoutLineItem } from './checkoutLineItem';

export interface CheckoutInput {
  /**
     * @minItems 1
     * @maxItems 20
     */
  items: CheckoutLineItem[];
}