# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: string;
  name: string;
  category: string;
  description: string;
  price: number;
  currency: string;
  image: string;
  material: string;
  featured: boolean;
  /** @nullable */
  badge: string | null;
  /** @nullable */
  stripePriceId: string | null;
}